Refactor MQTTScaleEvent handling: update weight type to float and adjust related processing logic

+++ stream-processor/src/main/java/com/eventdriven/healthcare/streamprocessor/util/ScaleFormatter.java
@@ -5,7 +5,7 @@ public class ScaleFormatter {
      * Accepts weights like "20"
      * and returns "weight minus the weight of the syringe (5gr)"
      */
-    public static int format(int raw) {
-        return raw - 4;
+    public static float format(float raw) {
+        return raw - 4f;
     }
 }
